requirements.txt
djangorestframework
model_bakery-1.15.0